feat: folder struct and ollama run
